+++ models/book_request.py
@@ -2,7 +2,7 @@ from peewee import Model, ForeignKeyField, CharField, DateTimeField
 from .db import db
 from .user import User
 
-class RequestBook(Model):
+class BookRequest(Model):
     user = ForeignKeyField(User, backref='requests')
     title = CharField()       # 申請したい本のタイトル
     author = CharField()      # 著者
