+++ templates/book_request_add.html
@@ -9,11 +9,11 @@
     <h1>新しい本のリクエスト</h1>
     <ul class="breadcrumb">
         <li><a href="{{ url_for('index') }}">HOME</a></li>
-        <li><a href="{{ url_for('request_book.list') }}">リクエスト一覧</a></li>
+        <li><a href="{{ url_for('book_request.list') }}">リクエスト一覧</a></li>
         <li>新規リクエスト</li>
     </ul>
     
-    <form action="{{ url_for('request_book.add') }}" method="post">
+    <form action="{{ url_for('book_request.add') }}" method="post">
         <label>ユーザー:
             <select name="user_id" required>
                 {% for user in users %}
