+++ templates/book_request_edit.html
@@ -10,11 +10,11 @@
 
     <ul class="breadcrumb">
         <li><a href="{{ url_for('index') }}">HOME</a></li>
-        <li><a href="{{ url_for('request_book.list') }}">リクエスト一覧</a></li>
+        <li><a href="{{ url_for('book_request.list') }}">リクエスト一覧</a></li>
         <li>編集</li>
     </ul>
 
-    <form action="{{ url_for('request_book.edit', request_id=req.id) }}" method="post">
+    <form action="{{ url_for('book_request.edit', request_id=req.id) }}" method="post">
         <label>ユーザー:
             <select name="user_id" required>
                 {% for user in users %}
