+++ templates/book_request_list.html
@@ -10,10 +10,10 @@
 
     <ul class="breadcrumb">
         <li><a href="{{ url_for('index') }}">HOME</a></li>
-        <li><a href="{{ url_for('request_book.list') }}">リクエスト一覧</a></li>
+        <li><a href="{{ url_for('book_request.list') }}">リクエスト一覧</a></li>
     </ul>
 
-    <a href="{{ url_for('request_book.add') }}">新規リクエスト追加</a>
+    <a href="{{ url_for('book_request.add') }}">新規リクエスト追加</a>
 
     <table>
         <thead>
@@ -29,7 +29,7 @@
         <tbody>
             {% for req in items %}
             <tr>
-                <td><a href="{{ url_for('request_book.edit', request_id=req.id) }}">編集</a></td>
+                <td><a href="{{ url_for('book_request.edit', request_id=req.id) }}">編集</a></td>
                 <td>{{ req.user.name }}</td>
                 <td>{{ req.title }}</td>
                 <td>{{ req.author }}</td>
