+++ templates/index.html
@@ -11,7 +11,7 @@
       <li><a href="{{ url_for('user.list') }}">ユーザー（顧客）</a></li>
       <li><a href="{{ url_for('product.list') }}">製品</a></li>
       <li><a href="{{ url_for('order.list') }}">注文</a></li>
-      <li><a href="{{ url_for('request_book.list') }}">本のリクエスト</a></li>
+      <li><a href="{{ url_for('book_request.list') }}">本のリクエスト</a></li>
     </ul>
 </body>
 </html>
